+++ src/themis/secure_message.h
@@ -37,61 +37,6 @@ extern "C"{
  */
 
 /**
- * @brief generate RSA key pair
- * @param [out] private_key buffer for private key to store. May be set to NULL for private key length determination
- * @param [in, out] private_key_length length of private_key
- * @param [out] public_key buffer for public key to store. May be set to NULL for public key length determination
- * @param [in, out] public_key_length length of public key
- * @return THEMIS_SUCCESS on success or THEMIS_FAIL on failure
- * @note If private_key==NULL or public_key==NULL or private_key_length is not enought for private key storage or public_key_length is not enought for public key storage then THEMIS_BUFFER_TOO_SMALL will return and private_key_length and public_key_length will store lengths of buffers needed for private key and public key store respectively
- */  
-themis_status_t themis_gen_rsa_key_pair(uint8_t* private_key,
-					size_t* private_key_length,
-					uint8_t* public_key,
-					size_t* public_key_length);
-
-/**
- * @brief generate EC key pair
- * @param [out] private_key buffer for private key to store. May be set to NULL for private key length determination
- * @param [in, out] private_key_length length of private_key
- * @param [out] public_key buffer for public key to store. May be set to NULL for public key length determination
- * @param [in, out] public_key_length length of public key
- * @return THEMIS_SUCCESS on success or THEMIS_FAIL on failure
- * @note If private_key==NULL or public_key==NULL or private_key_length is not enought for private key storage or public_key_length is not enought for public key storage then THEMIS_BUFFER_TOO_SMALL will return and private_key_length and public_key_length will store lengths of buffers needed for private key and public key store respectively
- */  
-themis_status_t themis_gen_ec_key_pair(uint8_t* private_key,
-				       size_t* private_key_length,
-				       uint8_t* public_key,
-				       size_t* public_key_length);
-
-enum themis_key_kind
-{
-  THEMIS_KEY_INVALID,
-  THEMIS_KEY_RSA_PRIVATE,
-  THEMIS_KEY_RSA_PUBLIC,
-  THEMIS_KEY_EC_PRIVATE,
-  THEMIS_KEY_EC_PUBLIC,
-};
-
-typedef enum themis_key_kind themis_key_kind_t;
-
-/**
- * @brief get Themis key kind
- * @param [in]  key     key buffer
- * @param [in]  length  length of key
- * @return corresponding key kind if the buffer contains a key, or THEMIS_KEY_INVALID otherwise
- */
-themis_key_kind_t themis_get_key_kind(const uint8_t* key, size_t length);
-
-/**
- * @brief validate a Themis key
- * @param [in]  key     key buffer to validate
- * @param [in]  length  length of key
- * @return THEMIS_SUCCESS if the buffer contains a valid Themis key, or an error code otherwise
- */
-themis_status_t themis_is_valid_key(const uint8_t* key, size_t length);
-
-/**
  * @brief encrypt message to secure message
  * @param [in]      private_key                 private key
  * @param [in]      private_key_length          length of private_key
