+++ src/themis/themis.h
@@ -33,6 +33,7 @@
 #define THEMIS_VERSION_TEXT "themis 0.9: "
 
 #include <themis/themis_error.h>
+#include <themis/secure_keygen.h>
 #include <themis/secure_message.h>
 #include <themis/secure_cell.h>
 #include <themis/secure_session.h>
